scripts/python/pyproject.toml
[project]
name = "pmk-sumatra-scripts"
version = "0.1.0"
description = "Python scripts for PMK Sumatra"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.0",
    "openai>=2.9.0",
    "pillow>=12.0.0",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
]

[tool.uv]
dev-dependencies = []
